Revert "ci(e2e): add rolling runtime history"
This reverts commit 436e5a3588524fbaeeebdc07b80063a4ca223ea0.
