+++ test/e2e/support/e2e-operations-workflow-boundary.test.ts
@@ -88,24 +88,6 @@ describe("E2E operations workflow boundary", () => {
     );
   });
 
-  it("retains a bounded runtime summary for future scheduled comparisons", () => {
-    const workflow = readE2eOperationsWorkflow();
-    const upload = workflow.jobs.scorecard.steps!.find(
-      (step) => step.name === "Upload E2E runtime summary",
-    );
-
-    expect(upload).toMatchObject({
-      if: "${{ always() && steps.scorecard.outcome == 'success' }}",
-      uses: "actions/upload-artifact@043fb46d1a93c77aae656e7c1c64a875d1fc6a0a",
-      with: {
-        name: "e2e-runtime-summary",
-        path: "${{ runner.temp }}/e2e-runtime-summary.json",
-        "if-no-files-found": "error",
-        "retention-days": 14,
-      },
-    });
-  });
-
   it("rejects controller protocol and PR validation drift", () => {
     const workflow = readE2eOperationsWorkflow();
     delete workflow.on?.workflow_dispatch?.inputs?.base_sha;
@@ -423,13 +405,9 @@ describe("E2E operations workflow boundary", () => {
         .fn()
         .mockReturnValue("## E2E Test Phase Runtime\n\n| Target | Slowest observed phase |"),
     };
-    const runtimeHistory = {
-      buildRuntimeHistory: vi.fn().mockResolvedValue("## E2E Nightly Runtime Trend"),
-    };
     const runtimeModules = new Map<string, unknown>([
       ["path", { join: (...parts: string[]) => parts.join("/") }],
       ["/workspace/scripts/audit-test-runtime.mts", runtimeAudit],
-      ["/workspace/scripts/scorecard/analyze-runtime-history.mts", runtimeHistory],
       ["/workspace/scripts/scorecard/coordinate-scorecard.mts", coordinator],
       ["/workspace/scripts/scorecard/analyze-trace-timing.mts", traceTiming],
       ["/workspace/scripts/scorecard/summarize-jobs.mts", scorecardJobs],
@@ -445,7 +423,6 @@ describe("E2E operations workflow boundary", () => {
         GITHUB_WORKSPACE: "/workspace",
         JOBS: "",
         RUNTIME_ARTIFACTS: "/runner/e2e-runtime-audit",
-        RUNTIME_SUMMARY_FILE: "/runner/e2e-runtime-summary.json",
         TARGETS: "",
       },
     };
@@ -468,11 +445,6 @@ describe("E2E operations workflow boundary", () => {
 
     expect(traceTiming.buildTraceTimingResult).toHaveBeenCalledWith({ github: {}, context, core });
     expect(runtimeAudit.auditTestRuntime).toHaveBeenCalledWith(["/runner/e2e-runtime-audit"]);
-    expect(runtimeHistory.buildRuntimeHistory).toHaveBeenCalledWith(
-      { github: {}, context, core },
-      [{ target: "full-e2e" }],
-      "/runner/e2e-runtime-summary.json",
-    );
     expect(runtimeAudit.auditTestRuntime.mock.invocationCallOrder[0]).toBeLessThan(
       traceTiming.buildTraceTimingResult.mock.invocationCallOrder[0],
     );
@@ -491,9 +463,7 @@ describe("E2E operations workflow boundary", () => {
       }),
     );
     expect(summary.addRaw).toHaveBeenCalledWith(
-      expect.stringMatching(
-        /### Onboard Performance Budget[\s\S]*## E2E Test Phase Runtime[\s\S]*## E2E Nightly Runtime Trend/u,
-      ),
+      expect.stringMatching(/### Onboard Performance Budget[\s\S]*## E2E Test Phase Runtime/u),
     );
     expect(summary.write).toHaveBeenCalledOnce();
     expect(setOutput).toHaveBeenCalledWith("scorecardData", expect.any(String));
@@ -518,13 +488,9 @@ describe("E2E operations workflow boundary", () => {
       }),
       formatRuntimeAuditSummary: vi.fn(),
     };
-    const runtimeHistory = {
-      buildRuntimeHistory: vi.fn().mockResolvedValue("## E2E Nightly Runtime Trend"),
-    };
     const runtimeModules = new Map<string, unknown>([
       ["path", { join: (...parts: string[]) => parts.join("/") }],
       ["/workspace/scripts/audit-test-runtime.mts", runtimeAudit],
-      ["/workspace/scripts/scorecard/analyze-runtime-history.mts", runtimeHistory],
       [
         "/workspace/scripts/scorecard/coordinate-scorecard.mts",
         {
@@ -561,7 +527,6 @@ describe("E2E operations workflow boundary", () => {
         GITHUB_WORKSPACE: "/workspace",
         JOBS: "",
         RUNTIME_ARTIFACTS: "/runner/e2e-runtime-audit",
-        RUNTIME_SUMMARY_FILE: "/runner/e2e-runtime-summary.json",
         TARGETS: "",
       },
     };
