+++ test/e2e/support/test-runtime-audit.test.ts
@@ -68,9 +68,6 @@ describe("test runtime audit", () => {
           p95Ms: 50_000,
           maxMs: 50_000,
           variabilityMs: 20_000,
-          passedRuns: 1,
-          failedRuns: 1,
-          skippedRuns: 0,
           slowestPhase: "inference",
           slowestPhaseMs: 40_000,
           slowestPhaseOutcome: "failed",
@@ -83,9 +80,6 @@ describe("test runtime audit", () => {
           p95Ms: 20_000,
           maxMs: 20_000,
           variabilityMs: 0,
-          passedRuns: 1,
-          failedRuns: 0,
-          skippedRuns: 0,
           slowestPhase: "sandbox",
           slowestPhaseMs: 15_000,
           slowestPhaseOutcome: "passed",
